[shared] Move endpoint context encoding and common json formats to layers
