+++ layers/src/main/scala/stasis/layers/security/tls/EndpointContext.scala
@@ -1,18 +1,24 @@
 package stasis.layers.security.tls
 
+import java.io.ByteArrayInputStream
+import java.io.ByteArrayOutputStream
 import java.io.FileInputStream
 import java.io.FileNotFoundException
 import java.security.KeyStore
 import java.security.SecureRandom
+import java.util.concurrent.ThreadLocalRandom
 
 import javax.net.ssl._
 
+import scala.jdk.CollectionConverters._
+import scala.util.Random
 import scala.util.Try
 
 import com.typesafe.{config => typesafe}
 import org.apache.pekko.http.scaladsl.ConnectionContext
 import org.apache.pekko.http.scaladsl.HttpsConnectionContext
 import org.apache.pekko.http.scaladsl.ServerBuilder
+import org.apache.pekko.util.ByteString
 
 final case class EndpointContext(
   config: EndpointContext.Config,
@@ -91,6 +97,92 @@ object EndpointContext {
         storeType = storeConfig.getString("type"),
         storePassword = storeConfig.getString("password")
       )
+  }
+
+  final case class Encoded(
+    enabled: Boolean,
+    protocol: String,
+    storeType: String,
+    temporaryStorePassword: String,
+    storeContent: ByteString
+  )
+
+  object Encoded {
+    final val TemporaryPasswordSize: Int = 32
+
+    final val DefaultProtocol: String = "TLS"
+    final val DefaultStoreType: String = "PKCS12"
+
+    def apply(config: typesafe.Config): Encoded =
+      if (config.getBoolean("enabled")) {
+        enabled(
+          protocol = config.getString("protocol"),
+          config = StoreConfig(config.getConfig("truststore"))
+        )
+      } else {
+        disabled()
+      }
+
+    def disabled(): Encoded =
+      Encoded(
+        enabled = false,
+        protocol = DefaultProtocol,
+        storeType = DefaultStoreType,
+        temporaryStorePassword = "",
+        storeContent = ByteString.empty
+      )
+
+    def enabled(
+      protocol: String,
+      config: StoreConfig
+    ): Encoded = {
+      val store = loadStore(config)
+
+      val rnd: Random = ThreadLocalRandom.current()
+      val password = rnd.alphanumeric.take(TemporaryPasswordSize).mkString
+
+      Encoded(
+        enabled = true,
+        protocol = protocol,
+        storeType = config.storeType,
+        temporaryStorePassword = password,
+        storeContent = encodeKeyStore(store, password, config.storeType)
+      )
+    }
+
+    @SuppressWarnings(Array("org.wartremover.warts.Null"))
+    def encodeKeyStore(store: KeyStore, password: String, storeType: String): ByteString = {
+      val rebuiltStore = KeyStore.getInstance(storeType)
+      rebuiltStore.load(None.orNull, None.orNull)
+
+      store.aliases().asScala.foreach {
+        case alias if store.isKeyEntry(alias) =>
+          store.getCertificateChain(alias).zipWithIndex.foreach { case (certificate, index) =>
+            rebuiltStore.setCertificateEntry(s"$alias-${index.toString}", certificate)
+          }
+
+        case alias =>
+          require(store.isCertificateEntry(alias), s"Expected certificate entry for alias [$alias]")
+          rebuiltStore.setCertificateEntry(alias, store.getCertificate(alias))
+      }
+
+      val content = new ByteArrayOutputStream()
+      rebuiltStore.store(content, password.toCharArray)
+
+      ByteString.fromArray(content.toByteArray)
+    }
+
+    def decodeKeyStore(content: ByteString, password: String, storeType: String): KeyStore = {
+      val store = KeyStore.getInstance(storeType)
+      store.load(new ByteArrayInputStream(content.toArray), password.toCharArray)
+
+      require(
+        store.aliases().asScala.forall(store.isCertificateEntry),
+        "Expected certificate entries only but one or more keys were encountered."
+      )
+
+      store
+    }
   }
 
   @SuppressWarnings(Array("org.wartremover.warts.Throw"))
