+++ layers/src/test/resources/application.conf
@@ -46,6 +46,16 @@ stasis {
             }
           }
 
+          context-enabled {
+            enabled = true
+            protocol = "TLS"
+            truststore {
+              path = "./core/src/test/resources/certs/localhost.p12"
+              type = "PKCS12"
+              password = ""
+            }
+          }
+
           context-disabled {
             enabled = false
             type = "client"
