+++ shared/src/test/resources/application.conf
@@ -1,22 +1,6 @@
 stasis {
   test {
     shared {
-      params {
-        context-enabled {
-          enabled = true
-          protocol = "TLS"
-          truststore {
-            path = "./core/src/test/resources/certs/localhost.p12"
-            type = "PKCS12"
-            password = ""
-          }
-        }
-
-        context-disabled {
-          enabled = false
-        }
-      }
-
       secrets {
         derivation {
           encryption {
