+++ shared/src/test/scala/stasis/test/specs/unit/shared/model/devices/DeviceBootstrapParametersSpec.scala
@@ -1,12 +1,7 @@
 package stasis.test.specs.unit.shared.model.devices
 
-import java.io.ByteArrayOutputStream
 import java.util.UUID
 
-import scala.jdk.CollectionConverters._
-
-import com.typesafe.{config => typesafe}
-import org.apache.pekko.util.ByteString
 import play.api.libs.json.Json
 
 import stasis.core.routing.Node
@@ -47,102 +42,6 @@ class DeviceBootstrapParametersSpec extends UnitSpec {
     deviceParams.serverApi.userSalt should be(user.salt)
   }
 
-  it should "support encoding and decoding a KeyStore to/from ByteString without private keys" in {
-    val config = EndpointContext.StoreConfig(
-      storePath = "./core/src/test/resources/certs/localhost.p12",
-      storeType = "PKCS12",
-      storePassword = ""
-    )
-
-    val password = "test-password"
-
-    val original = EndpointContext.loadStore(config)
-
-    val encoded = DeviceBootstrapParameters.Context.encodeKeyStore(original, password, config.storeType)
-    val decoded = DeviceBootstrapParameters.Context.decodeKeyStore(encoded, password, config.storeType)
-
-    val encodedWithCertsOnly = DeviceBootstrapParameters.Context.encodeKeyStore(decoded, password, config.storeType)
-    val decodedWithCertsOnly = DeviceBootstrapParameters.Context.decodeKeyStore(encodedWithCertsOnly, password, config.storeType)
-
-    original.size should be > 0
-    original.aliases().asScala.toList.exists(original.isKeyEntry)
-
-    decoded.size should be > 0
-    decoded.aliases().asScala.toList.forall(original.isCertificateEntry)
-
-    decodedWithCertsOnly.size should be > 0
-    decodedWithCertsOnly.aliases().asScala.toList.forall(original.isCertificateEntry)
-  }
-
-  it should "fail to decode a KeyStore with private keys" in {
-    val config = EndpointContext.StoreConfig(
-      storePath = "./core/src/test/resources/certs/localhost.p12",
-      storeType = "PKCS12",
-      storePassword = ""
-    )
-
-    val password = "test-password"
-
-    val original = EndpointContext.loadStore(config)
-
-    val content = new ByteArrayOutputStream()
-    original.store(content, password.toCharArray)
-
-    val encoded = ByteString.fromArray(content.toByteArray)
-
-    an[IllegalArgumentException] should be thrownBy {
-      DeviceBootstrapParameters.Context.decodeKeyStore(encoded, password, config.storeType)
-    }
-  }
-
-  it should "support creating enabled contexts" in {
-    val config = EndpointContext.StoreConfig(
-      storePath = "./core/src/test/resources/certs/localhost.p12",
-      storeType = "PKCS12",
-      storePassword = ""
-    )
-
-    val actualContext = DeviceBootstrapParameters.Context.enabled(protocol = "TLS", config = config)
-
-    actualContext.enabled should be(true)
-    actualContext.protocol should be("TLS")
-    actualContext.storeType should be(config.storeType)
-    actualContext.temporaryStorePassword.length should be(DeviceBootstrapParameters.Context.TemporaryPasswordSize)
-    actualContext.storeContent should not be empty
-  }
-
-  it should "support creating disabled contexts" in {
-    DeviceBootstrapParameters.Context.disabled() should be(
-      DeviceBootstrapParameters.Context(
-        enabled = false,
-        protocol = "TLS",
-        storeType = "PKCS12",
-        temporaryStorePassword = "",
-        storeContent = ByteString.empty
-      )
-    )
-  }
-
-  it should "support creating contexts from config" in {
-    val config = typesafe.ConfigFactory.load().getConfig("stasis.test.shared.params")
-
-    val enabledContext = DeviceBootstrapParameters.Context(config.getConfig("context-enabled"))
-
-    enabledContext.enabled should be(true)
-    enabledContext.protocol should be("TLS")
-    enabledContext.storeType should be("PKCS12")
-    enabledContext.temporaryStorePassword.length should be(DeviceBootstrapParameters.Context.TemporaryPasswordSize)
-    enabledContext.storeContent should not be empty
-
-    val disabledContext = DeviceBootstrapParameters.Context(config.getConfig("context-disabled"))
-
-    disabledContext.enabled should be(false)
-    disabledContext.protocol should be(DeviceBootstrapParameters.Context.DefaultProtocol)
-    disabledContext.storeType should be(DeviceBootstrapParameters.Context.DefaultStoreType)
-    disabledContext.temporaryStorePassword should be(empty)
-    disabledContext.storeContent should be(empty)
-  }
-
   private val baseParams = DeviceBootstrapParameters(
     authentication = DeviceBootstrapParameters.Authentication(
       tokenEndpoint = "http://localhost:1234",
@@ -153,19 +52,19 @@ class DeviceBootstrapParametersSpec extends UnitSpec {
         api = "urn:stasis:identity:audience:server-api",
         core = s"urn:stasis:identity:audience:${Node.generateId().toString}"
       ),
-      context = DeviceBootstrapParameters.Context.disabled()
+      context = EndpointContext.Encoded.disabled()
     ),
     serverApi = DeviceBootstrapParameters.ServerApi(
       url = "http://localhost:5678",
       user = "",
       userSalt = "",
       device = "",
-      context = DeviceBootstrapParameters.Context.disabled()
+      context = EndpointContext.Encoded.disabled()
     ),
     serverCore = DeviceBootstrapParameters.ServerCore(
       address = "http://localhost:5679",
       nodeId = "",
-      context = DeviceBootstrapParameters.Context.disabled()
+      context = EndpointContext.Encoded.disabled()
     ),
     secrets = SecretsConfig(
       derivation = SecretsConfig.Derivation(
